+++ admin/admin.php
@@ -1,6 +1,6 @@
 <?php
 /*
-Version:     6.5
+Version:     6.7
 Date:        21/12/25
 Name:        admin.php
 Purpose:     Site control panel
@@ -176,7 +176,7 @@ if ($admin !== 1) :
 endif;
 
 //Get date for update form
-$dateObject = new DateYMD();
+$dateObject = new \MTG\Core\DateYMD();
 $date = $dateObject->getToday();
 
 $scryAction = filter_input(INPUT_POST, 'scryfalljson_action', FILTER_UNSAFE_RAW);
@@ -538,7 +538,7 @@ if ($configAction === 'start_reauth') :
     $configAuthRequested = true;
 elseif ($configAction === 'reauth_submit') :
     $reauthPassword = filter_input(INPUT_POST, 'config_password', FILTER_UNSAFE_RAW);
-    $passwordCheck = new PasswordCheck($db, $logfile, $siteTitle);
+    $passwordCheck = new \MTG\Auth\PasswordCheck($db, $logfile, $siteTitle);
     $reauthResult = $passwordCheck->validatePassword($userEmail, $reauthPassword);
     if ($reauthResult === 10) :
         $_SESSION['config_edit_expires'] = time() + $configAuthWindowSeconds;