+++ classes/deckmanager.class.php
@@ -12,9 +12,7 @@ To do:       -
 */
 
 // phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespace, PSR1.Files.SideEffects.FoundWithSymbols
-if (__FILE__ == $_SERVER['PHP_SELF']) :
-    die('Direct access prohibited');
-endif;
+
 
 class DeckManager
 {